♻ extract persistence
